Crate quantity

The Quantity Programming Language

A type-driven, mid-level programming language meant for scientific computation, mathematics, and simulation

All links on this page will redirect you to the official documentation for this project

Every decision made in the design of this language is based off the following axioms:

Behavior is defined

Boilerplate is wrong

Limitations are few

Values have meaning

Speed is implicit

Functionality is composed

Organization is important

Documentation is powerful

Page	Description

Docs	Documentation related to the language itself

Interface	Documentation about how the interpreter can be run

Interpreting	Documentation about the interpreter itself

Parsing	Documentation about how the AST is generated

Scanning	Documentation about how the source code is parsed into tokens

Concept	Description

Explicit	Quantity code shall have no undefined behavior

Beautiful	Quantity code shall be easy to read and understand

Functional	Quantity code shall be purely functional yet organized

Expressive	Quantity code shall require minimal boilerplate to be productive

Modules

docs 🔒

This module contains documentation related to the language itself