Crate quantity

The Quantity Programming Language

A high-performance statically typed programming language meant for scientific computation, simulation, and system control.

Quickly write highly performant code with complete confidence of its correctness!

qty length meters|meter|m
qty area meters^2|meter^2|m^2 (length*length)

rect_area = area|length l, length w| {
    l*w
}

print rect_area( 12m, 2m )

Please note that all links on this page will redirect you to the official documentation for this project

The Seven Laws of Quantity

Listed below are the laws that describe and define this language, in order of their significance. Clicking any on any of them will take you to a more detailed description of their meaning.

ğŸª Law of Reality: the world or the state of things as they actually exist

ğŸ Law of Amenity: an inessential, yet desirable item that provides added comfort

ğŸ“ Law of Brevity: concise and exact use of words in writing or speech

ğŸª¢ Law of Tensity: the state of being rigid or inflexible

â˜€ï¸ Law of Clarity: quality or state of being clear

â˜¯ï¸ Law of Duality: an instance of opposition or contrast between two concepts

ğŸ‡ Law of Rabbity: resembling, characteristic of, or containing rabbits

Modules

docs ğŸ”’

This module contains documentation related to the language itself