Crate quantity

The Quantity Programming Language

A type-driven, mid-level programming language meant for scientific computation and simulation

The main design points driving the creation of this programming language are:

Concept	Description

Explicit	Quantity code shall have no undefined behavior

Beautiful	Quantity code shall be easy to read and understand

Functional	Quantity code shall be purely functional yet organized

Expressive	Quantity code shall require minimal boilerplate to be productive

The following table will link to the official documentation for this project

Page	Description

Docs	The documentation related to the language itself

Interface	Documentation about how the interpreter can be run

Interpreting	Documentation about the interpreter itself

Parsing	Documentation about how the AST is generated

Scanning	Documentation about how the source code is parsed into tokens

Quantity Code Examples

The following snippets of code show what is (or will be) possible using the quantity programming language

Quantities

Quantities are the backbone of this language, so it only makes sense to start this guide by making some quantities.

// Declaring new quantities for the program
qty length
qty area (length*length)
qty volume (length*area)
qty time
qty rate (length/time)

As you can see, a quantity is basically a system of measurement. They are almost like different realms of numbers. You can add and subtract two lengths together, but you cannot add a length to an area or volume. We can, however, move between quantities using multiplication and division. For example, two lengths multiplied together produce an area quantity.

In simple terms, you can think of the quantities as generic data types that can be satisfied by any of their member units.

Units

Units and quantities go hand-in-hand. Think of quantities as groupings of units. The length quantity contains meters, feet, lightyears, etc. These can freely be added, subtracted, or even converted between each other.

// Declaring new units for the program
length meter|meters|m
length foot|feet|ft = 0.3048m

The first example line above is the declaration of a base unit. It does not need any conversions. Afterwards, the feet unit is declared in relation to the meter unit.

Number Literals

Now that we have quantities and units defined, we can start to create number literals with them.

For example, if 34 meters ever shows up in the code, it will generate a value of 34 and attach it to the meter ‘type’.

The following code block will show the end value of various expressions:

45 meters           // 45 meters
3.281ft             // 1 meter
1meter * 3.281 feet // 1 meter*meter

Name Literals

Naming conventions are much more lax in Quantity than perhaps any other programming language. A name can contain any of the following characters:

Category	Acceptable Characters	Start of Name

Alphabetic	a-z A-Z	Yes

Numeric	0-9	No

Symbolic	- + * / etc	Yes

Emojis	all	Yes

The Start of Name column states whether or not the name can begin with that category of text.

Functions

Since Quantity is a functional language, it has first-class support for functions, which end up being extremely similar to closures in Rust. These functions also include rich support for generic types based on quantities.

area rect_area = |length len, length wid| len * wid
area tri_area = |length base, length height| base * height

There are three main ways to call functions:

Traditional Syntax
The traditional approach is the same across virtually all programming languages. It involves using the function name and passing in every argument.

rect_area(45ft, 10m) // Returns an area value which carries the ft*m flag

Method Syntax
While Quantity is not object oriented, we can still benefit from the easier organization that object oriented languages sometimes provide. To use this syntax, the first argument is treated as an object, the function name is treated as a member, and the other arguments are passed into the function as normal.

45ft.rect_area(10m) // Returns an area value which carries the ft*m flag

Operator Syntax
Any function that takes two arguments can be used as an operator. This mostly makes sense with functions that perform actual operations, but this language feature can be used by any dual-argument function.

45ft rect_area 10m // Returns an area value which carries the ft*m flag

Modules

docs 🔒

This module contains documentation related to the language itself

interface 🔒

The interface module contains logic that allows the interpreter to run both as an interpreter and as a REPL. This module is not about the language so much as it is about how the language is read.

interpreting 🔒

parsing 🔒

scanning 🔒

Functions

main 🔒

run